Cite

Documentation for Cite.

Cite.collect_citations

Cite.get_citations

Cite.get_tool_citation

Cite.collect_citations — Method

collect_citations(only_direct::Bool)

Collect the citations from all the dependencies in the current environment.

Cite.get_citations — Method

get_citations(; only_direct=false, filename="julia_citations.bib")

This function will create a .bib file with all the citations collected form the CITATION.bib files corresponding to the dependecies of the current active environment. Use filename to change the name of the file. To include just the direct dependencies use only_direct=true

Cite.get_tool_citation — Function

get_tool_citation(io::IO=stdout; jl = true, texttt = false, copy = true, cite_commands=Dict{String,String}(), filename="julia_citations.bib")

Print a sentence describing the packages used in the current environment. If you only want to consider the direct dependencies, you can set only_direct=true. The sentence is automatically copied to the clipboard(you can avoid this by using copy=false). The package names have the .jl ending by default. You can ommit it with jl=false. Package names can be wrapped in texttt by setting texttt=true and you can also customize the cite command used for each package by using cite_commands=Dict("PackageName"=>"custom_cite"). The filename of the .bib file can be passed via the filename keyword.